Revert "Could size_t be too small?"
This reverts commit b681feab

+++ src/squashdelta.cxx
@@ -33,17 +33,15 @@ extern "C"
 
 struct compressed_block
 {
-	// Host endianness
-	uint64_t offset;
-	uint32_t length;
-	uint32_t uncompressed_length;
+	size_t offset;
+	size_t length;
+	size_t uncompressed_length;
 	uint32_t hash;
 };
 
 #pragma pack(push, 1)
 struct serialized_compressed_block
 {
-	// Shall be big-endian
 	uint64_t offset;
 	uint32_t length;
 	uint32_t uncompressed_length;
